Bug 1669181 - Part 42: Split InlineScriptTree and BytecodeSite from "jit/CompileInfo.h". r=jandem
"jit/MIR.h" needs to access `InlineScriptTree` and `BytecodeSite`. By splitting
both classes from "jit/CompileInfo.h" we can avoid pulling in the additional
headers from "jit/CompileInfo.h" into "jit/MIR.h". This is useful because
"jit/MIR.h" is (transitively) included in many other headers.

Differential Revision: https://phabricator.services.mozilla.com/D92732
