+++ js/src/jit/CompileInfo-inl.h
@@ -8,7 +8,6 @@
 #define jit_CompileInfo_inl_h
 
 #include "jit/CompileInfo.h"
-#include "jit/JitAllocPolicy.h"
 
 #include "vm/JSScript-inl.h"
 
@@ -21,49 +20,6 @@ inline RegExpObject* CompileInfo::getRegExp(jsbytecode* pc) const {
 
 inline JSFunction* CompileInfo::getFunction(jsbytecode* pc) const {
   return script_->getFunction(pc);
-}
-
-InlineScriptTree* InlineScriptTree::New(TempAllocator* allocator,
-                                        InlineScriptTree* callerTree,
-                                        jsbytecode* callerPc,
-                                        JSScript* script) {
-  MOZ_ASSERT_IF(!callerTree, !callerPc);
-  MOZ_ASSERT_IF(callerTree, callerTree->script()->containsPC(callerPc));
-
-  // Allocate a new InlineScriptTree
-  void* treeMem = allocator->allocate(sizeof(InlineScriptTree));
-  if (!treeMem) {
-    return nullptr;
-  }
-
-  // Initialize it.
-  return new (treeMem) InlineScriptTree(callerTree, callerPc, script);
-}
-
-InlineScriptTree* InlineScriptTree::addCallee(TempAllocator* allocator,
-                                              jsbytecode* callerPc,
-                                              JSScript* calleeScript) {
-  MOZ_ASSERT(script_ && script_->containsPC(callerPc));
-  InlineScriptTree* calleeTree = New(allocator, this, callerPc, calleeScript);
-  if (!calleeTree) {
-    return nullptr;
-  }
-
-  calleeTree->nextCallee_ = children_;
-  children_ = calleeTree;
-  return calleeTree;
-}
-
-void InlineScriptTree::removeCallee(InlineScriptTree* callee) {
-  InlineScriptTree** prevPtr = &children_;
-  for (InlineScriptTree* child = children_; child; child = child->nextCallee_) {
-    if (child == callee) {
-      *prevPtr = child->nextCallee_;
-      return;
-    }
-    prevPtr = &child->nextCallee_;
-  }
-  MOZ_CRASH("Callee not found");
 }
 
 static inline const char* AnalysisModeString(AnalysisMode mode) {
