+++ js/src/jit/IonBuilder.h
@@ -16,6 +16,8 @@
 
 #include "builtin/TypedObject.h"
 #include "jit/BaselineInspector.h"
+#include "jit/CompileInfo.h"
+#include "jit/InlineScriptTree.h"
 #include "jit/IonAnalysis.h"
 #include "jit/IonOptimizationLevels.h"
 #include "jit/MIR.h"
