+++ js/src/jit/MIR.h
@@ -23,9 +23,9 @@
 #include "builtin/ModuleObject.h"
 #include "jit/AtomicOp.h"
 #include "jit/BaselineIC.h"
-#include "jit/CompileInfo.h"
 #include "jit/FixedList.h"
 #include "jit/InlineList.h"
+#include "jit/InlineScriptTree.h"
 #include "jit/JitAllocPolicy.h"
 #include "jit/MOpcodesGenerated.h"
 #include "jit/TIOracle.h"
