+++ js/src/jit/MIRGraph.cpp
@@ -6,6 +6,8 @@
 
 #include "jit/MIRGraph.h"
 
+#include "jit/CompileInfo.h"
+#include "jit/InlineScriptTree.h"
 #include "jit/Ion.h"
 #include "jit/JitSpewer.h"
 #include "jit/MIR.h"
