+++ js/src/jit/Recover.cpp
@@ -12,6 +12,7 @@
 #include "builtin/RegExp.h"
 #include "builtin/String.h"
 #include "builtin/TypedObject.h"
+#include "jit/CompileInfo.h"
 #include "jit/Ion.h"
 #include "jit/JitSpewer.h"
 #include "jit/JSJitFrameIter.h"
