+++ js/src/jit/arm64/CodeGenerator-arm64.cpp
@@ -11,6 +11,7 @@
 #include "jsnum.h"
 
 #include "jit/CodeGenerator.h"
+#include "jit/InlineScriptTree.h"
 #include "jit/JitFrames.h"
 #include "jit/JitRuntime.h"
 #include "jit/MIR.h"
