+++ js/src/jit/shared/CodeGenerator-shared.h
@@ -11,6 +11,7 @@
 
 #include <utility>
 
+#include "jit/InlineScriptTree.h"
 #include "jit/JitcodeMap.h"
 #include "jit/JitFrames.h"
 #include "jit/LIR.h"
