+++ js/src/jit/x86-shared/CodeGenerator-x86-shared.cpp
@@ -12,6 +12,7 @@
 #include "jsmath.h"
 
 #include "jit/CodeGenerator.h"
+#include "jit/InlineScriptTree.h"
 #include "jit/JitFrames.h"
 #include "jit/JitRuntime.h"
 #include "jit/RangeAnalysis.h"
